Working state app
I got the state app working! It's not perfect and needs several features but works. Also added shape files for state membership and crop data. City shape files are too big to be hosted on Github without LFS so for the city app I plan on joining the localharvest production data to the tigris shapefile inside the app,.